review/requirements.txt
Django==3.2
djangorestframework==3.11.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0